FIX - Date Publication line Height (#209)
* fix lineHeight and update snapshots

* refactoring tests

* remove shared
